refactor update checker to use AppState
