+++ modules/updatechecker/update_checker.go
@@ -2,29 +2,28 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package models
+package updatechecker
 
 import (
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"net/http"
 
-	"code.gitea.io/gitea/models/db"
+	"code.gitea.io/gitea/modules/appstate"
 	"code.gitea.io/gitea/modules/proxy"
 	"code.gitea.io/gitea/modules/setting"
 
 	"github.com/hashicorp/go-version"
 )
 
-// RemoteVersion stores the remote version from the JSON endpoint
-type RemoteVersion struct {
-	ID      int64  `xorm:"pk autoincr"`
-	Version string `xorm:"VARCHAR(50)"`
+// CheckerState stores the remote version from the JSON endpoint
+type CheckerState struct {
+	LatestVersion string
 }
 
-func init() {
-	db.RegisterModel(new(RemoteVersion))
+// Name returns the name of the state item for update checker
+func (r *CheckerState) Name() string {
+	return "update-checker"
 }
 
 // GiteaUpdateChecker returns error when new version of Gitea is available
@@ -49,60 +48,33 @@ func GiteaUpdateChecker(httpEndpoint string) error {
 		return err
 	}
 
-	type v struct {
+	type respType struct {
 		Latest struct {
 			Version string `json:"version"`
 		} `json:"latest"`
 	}
-	ver := v{}
-	err = json.Unmarshal(body, &ver)
+	respData := respType{}
+	err = json.Unmarshal(body, &respData)
 	if err != nil {
 		return err
 	}
 
-	return UpdateRemoteVersion(ver.Latest.Version)
+	return UpdateRemoteVersion(respData.Latest.Version)
 
 }
 
 // UpdateRemoteVersion updates the latest available version of Gitea
 func UpdateRemoteVersion(version string) (err error) {
-	sess := db.NewSession(db.DefaultContext)
-	defer sess.Close()
-	if err = sess.Begin(); err != nil {
-		return err
-	}
-
-	currentVersion := &RemoteVersion{ID: 1}
-	has, err := sess.Get(currentVersion)
-	if err != nil {
-		return fmt.Errorf("get: %v", err)
-	} else if !has {
-		currentVersion.ID = 1
-		currentVersion.Version = version
-
-		if _, err = sess.InsertOne(currentVersion); err != nil {
-			return fmt.Errorf("insert: %v", err)
-		}
-		return nil
-	}
-
-	if _, err = sess.Update(&RemoteVersion{ID: 1, Version: version}); err != nil {
-		return err
-	}
-
-	return sess.Commit()
+	return appstate.AppState.Set(&CheckerState{LatestVersion: version})
 }
 
 // GetRemoteVersion returns the current remote version (or currently installed verson if fail to fetch from DB)
 func GetRemoteVersion() string {
-	e := db.GetEngine(db.DefaultContext)
-	v := &RemoteVersion{ID: 1}
-	_, err := e.Get(&v)
-	if err != nil {
-		// return current version if fail to fetch from DB
-		return setting.AppVer
+	item := new(CheckerState)
+	if err := appstate.AppState.Get(item); err != nil {
+		return ""
 	}
-	return v.Version
+	return item.LatestVersion
 }
 
 // GetNeedUpdate returns true whether a newer version of Gitea is available
@@ -112,7 +84,12 @@ func GetNeedUpdate() bool {
 		// return false to fail silently
 		return false
 	}
-	remoteVer, err := version.NewVersion(GetRemoteVersion())
+	remoteVerStr := GetRemoteVersion()
+	if remoteVerStr == "" {
+		// no remote version is known
+		return false
+	}
+	remoteVer, err := version.NewVersion(remoteVerStr)
 	if err != nil {
 		// return false to fail silently
 		return false
